+++ src/lib/pipeline/BalancedPortAllocator.cc
@@ -9,12 +9,12 @@ BalancedPortAllocator::BalancedPortAllocator(
     const std::vector<std::vector<uint16_t>>& portArrangement)
     : weights(portArrangement.size(), 0) {}
 
-uint8_t BalancedPortAllocator::allocate(const std::vector<uint8_t>& ports) {
+uint16_t BalancedPortAllocator::allocate(const std::vector<uint16_t>& ports) {
   assert(ports.size() &&
          "No supported ports supplied; cannot allocate from a empty set");
   bool foundPort = false;
   uint16_t bestWeight = 0xFFFF;
-  uint8_t bestPort = 0;
+  uint16_t bestPort = 0;
   for (const auto& portIndex : ports) {
     // Search for the lowest-weighted port available
     if (!foundPort || weights[portIndex] < bestWeight) {
@@ -31,11 +31,11 @@ uint8_t BalancedPortAllocator::allocate(const std::vector<uint8_t>& ports) {
   return bestPort;
 }
 
-void BalancedPortAllocator::issued(uint8_t port) {
+void BalancedPortAllocator::issued(uint16_t port) {
   assert(weights[port] > 0);
   weights[port]--;
 }
-void BalancedPortAllocator::deallocate(uint8_t port) { issued(port); };
+void BalancedPortAllocator::deallocate(uint16_t port) { issued(port); };
 
 void BalancedPortAllocator::setRSSizeGetter(
     std::function<void(std::vector<uint64_t>&)> rsSizes) {
